feat: Implement authentication models and repository
- Added AuthResponseModel, LoginRequestBodyModel, RegisterRequestBodyModel, and UserData models.
- Created AuthRepo for handling login and registration logic.
- Removed obsolete login and register data sources and repositories.
- Introduced LoginCubit and RegisterCubit for managing login and registration states.
- Developed UI components for login and registration screens, including forms and buttons.
- Integrated social media sign-up options.
- Updated dependencies in pubspec.yaml and pubspec.lock.
