+++ lib/core/config/routing/app_router.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 import '../../../features/addresses/presentation/ui/addresses_screen.dart';
-import '../../../features/auth/login/logic/login_cubit.dart';
-import '../../../features/auth/login/presentation/login_screen.dart';
-import '../../../features/auth/register/logic/register_cubit.dart';
-import '../../../features/auth/register/presentation/register_screen.dart';
+import '../../../features/auth/presentation/logic/login_cubit/login_cubit.dart';
+import '../../../features/auth/presentation/logic/register_cubit/register_cubit.dart';
+import '../../../features/auth/presentation/ui/login/login_screen.dart';
+import '../../../features/auth/presentation/ui/register/register_screen.dart';
 import '../../../features/cars/presentation/logic/cars_cubit.dart';
 import '../../../features/cars/presentation/ui/car_brands_screen.dart';
 import '../../../features/cars/presentation/ui/cars_screen.dart';
@@ -14,7 +14,7 @@ import '../../../features/checkout/checkout_screen.dart';
 import '../../../features/favorites/presentation/logic/favorites_cubit.dart';
 import '../../../features/garage/presentation/logic/garage_cubit.dart';
 import '../../../features/layout/logic/layout_cubit.dart';
-import '../../../features/layout/presentation/layout_screen.dart';
+import '../../../features/layout/ui/layout_screen.dart';
 import '../../../features/on_boarding/on_boarding_screen.dart';
 import '../../../features/orders/presentation/ui/orders_screen.dart';
 import '../../../features/payment_methods/presentation/ui/payments_screen.dart';
