+++ lib/core/di/dependency_injection.dart
@@ -1,12 +1,10 @@
 import 'package:dio/dio.dart';
 import 'package:get_it/get_it.dart';
 
-import '../../features/auth/login/data/data_sources/login_remote_data_source.dart';
-import '../../features/auth/login/data/repos/login_repo.dart';
-import '../../features/auth/login/logic/login_cubit.dart';
-import '../../features/auth/register/data/data_sources/register_remote_data_source.dart';
-import '../../features/auth/register/data/repos/register_repo.dart';
-import '../../features/auth/register/logic/register_cubit.dart';
+import '../../features/auth/data/data_sources/auth_remote_data_source.dart';
+import '../../features/auth/data/repos/auth_repo.dart';
+import '../../features/auth/presentation/logic/login_cubit/login_cubit.dart';
+import '../../features/auth/presentation/logic/register_cubit/register_cubit.dart';
 import '../../features/cars/data/data_sources/car_brands_remote_data_source.dart';
 import '../../features/cars/data/data_sources/car_brands_remote_data_source_impl.dart';
 import '../../features/cars/data/repos_impl/car_brands_repo_impl.dart';
@@ -74,6 +72,7 @@ import '../../features/search/domain/use_cases/search_use_case.dart';
 import '../../features/search/presentation/logic/search_cubit.dart';
 import '../caching/hive_manager.dart';
 import '../helpers/app_bloc_observer.dart';
+import '../networking/connectivity_service.dart';
 import '../networking/crud_manager.dart';
 import '../networking/dio/dio_factory.dart';
 
@@ -83,6 +82,8 @@ Future<void> initGetIt() async {
   // Dio & ApiService & HiveManager
   Dio freeDio = DioFactory.getFreeDio();
   Dio tokenDio = DioFactory.getTokenDio();
+
+  getIt.registerLazySingleton<ConnectivityService>(() => ConnectivityService());
 
   getIt.registerLazySingleton<CrudManager>(
     () => CrudManager.getInstance(
@@ -98,19 +99,13 @@ Future<void> initGetIt() async {
   // BLoc observer
   getIt.registerSingleton<AppBlocObserver>(AppBlocObserver());
 
-  // Register
-  getIt.registerFactory<RegisterRemoteDataSource>(
-    () => RegisterRemoteDataSource(getIt()),
+  // auth
+  getIt.registerFactory<AuthRemoteDataSource>(
+    () => AuthRemoteDataSource(getIt()),
   );
-  getIt.registerFactory<RegisterRepo>(() => RegisterRepo(getIt()));
-  getIt.registerFactory<RegisterCubit>(() => RegisterCubit(getIt()));
-
-  // login
-  getIt.registerFactory<LoginRemoteDataSource>(
-    () => LoginRemoteDataSource(getIt()),
-  );
-  getIt.registerFactory<LoginRepo>(() => LoginRepo(getIt()));
+  getIt.registerFactory<AuthRepo>(() => AuthRepo(getIt()));
   getIt.registerFactory<LoginCubit>(() => LoginCubit(getIt()));
+  getIt.registerFactory<RegisterCubit>(() => RegisterCubit(getIt()));
 
   // home
   getIt.registerFactory<HomeDataSources>(
