+++ lib/core/errors/api_error_model.dart
@@ -1,0 +1,11 @@
+class ApiErrorModel {
+  int? statusCode;
+  String? message;
+  dynamic data;
+
+  ApiErrorModel({
+    this.statusCode,
+    this.message,
+    this.data,
+  });
+}
