+++ lib/core/errors/local_error_exception.dart
@@ -1,0 +1,5 @@
+class LocalErrorException implements Exception {
+  final String? message;
+
+  LocalErrorException({this.message});
+}
