+++ lib/core/networking/dio/dio_factory.dart
@@ -51,17 +51,14 @@ class DioFactory {
   }
 
   static void _addFreeDioInterceptors() {
-    _freeDio!.interceptors.addAll(
-      [
-        !kReleaseMode
-            ? PrettyDioLogger(
-                requestHeader: true,
-                requestBody: true,
-                responseBody: true,
-                responseHeader: false,
-              )
-            : const Interceptor(),
-      ],
+    _freeDio!.interceptors.add(
+      PrettyDioLogger(
+        requestHeader: true,
+        requestBody: true,
+        responseBody: true,
+        responseHeader: false,
+        enabled: !kReleaseMode,
+      ),
     );
   }
 
@@ -69,14 +66,13 @@ class DioFactory {
     _tokenDio!.interceptors.addAll(
       [
         TokenInterceptor(),
-        !kReleaseMode
-            ? PrettyDioLogger(
-                requestHeader: true,
-                requestBody: true,
-                responseBody: true,
-                responseHeader: false,
-              )
-            : const Interceptor(),
+        PrettyDioLogger(
+          requestHeader: true,
+          requestBody: true,
+          responseBody: true,
+          responseHeader: false,
+          enabled: !kReleaseMode,
+        ),
       ],
     );
   }
