--- lib/features/auth/login/logic/login_state.dart
@@ -1,20 +1,0 @@
-part of 'login_cubit.dart';
-
-@immutable
-sealed class LoginState {}
-
-final class InitialState extends LoginState {}
-
-final class LoadingState extends LoginState {}
-
-final class SuccessState extends LoginState {
-  final String message;
-
-  SuccessState(this.message);
-}
-
-final class ErrorState extends LoginState {
-  final String message;
-
-  ErrorState(this.message);
-}
