+++ lib/features/auth/presentation/logic/login_cubit/login_state.dart
@@ -1,0 +1,16 @@
+part of 'login_cubit.dart';
+
+@immutable
+sealed class LoginState {}
+
+final class InitialState extends LoginState {}
+
+final class LoadingState extends LoginState {}
+
+final class SuccessState extends LoginState {}
+
+final class ErrorState extends LoginState {
+  final ApiErrorModel error;
+
+  ErrorState(this.error);
+}
