+++ pubspec.yaml
@@ -13,7 +13,7 @@ dependencies:
   cached_network_image: ^3.4.1
   carousel_slider_plus: ^6.0.0+1
   colorful_safe_area: ^1.0.0
-  dartz: ^0.10.1
+  connectivity_plus: ^6.1.4
   dio: ^5.4.3+1
   equatable: ^2.0.5
   flutter:
@@ -33,8 +33,7 @@ dependencies:
   path_provider: ^2.1.5
   persistent_bottom_nav_bar: ^6.2.1
   popover: ^0.3.0+1
-  pretty_dio_logger: ^1.3.1
-  # shared_preferences: ^2.3.0
+  pretty_dio_logger: ^1.4.0
   shimmer: ^3.0.0
 
 dev_dependencies:
